# backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: Good Doggy CRM API
  version: 0.1.0
  description: Contrato de la API del backend. Fuente para el cliente TypeScript (orval).
paths:
...
  /api/v1/o/{org_slug}/me/:
    get:
      operationId: me_context
      description: |-
        Quién es el usuario en esta organización y qué puede hacer. Informa: no autoriza. La
        interfaz lo usa para construir la navegación; cada ruta sigue comprobando su permiso.
      parameters:
      - in: path
        name: org_slug
        schema:
          type: string
        required: true
      tags:
      - me
      security:
      - sessionCookie: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfContext'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/v1/o/{org_slug}/members/:
    get:
      operationId: members_list
      description: |-
        Quién pertenece a la organización, en cualquier estado, con qué roles y en qué sucursal.
        Paginado por orden de alta (ADR-016).
      parameters:
      - name: cursor
        in: query
        required: false
        description: El `next` de la página anterior. Sin él, la primera página.
        schema:
          type: string
      - name: limit
        in: query
        required: false
        description: Filas por página.
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - in: path
        name: org_slug
        schema:
          type: string
        required: true
      tags:
      - members
      security:
      - sessionCookie: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMemberList'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
  /api/v1/o/{org_slug}/members/{membership_id}/roles/{role_id}/:
    put:
      operationId: members_roles_assign
      description: 'Asigna el rol al miembro. Repetirlo no cambia nada. 403: sin `users.manage`,
        uno mismo, o un rol que el actor no cubre. 404: el miembro o el rol no son
        de la organización. 409 `LAST_OWNER`: la organización no tiene rol Owner y
        no admite ningún cambio.'
      parameters:
      - in: path
        name: membership_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: org_slug
        schema:
          type: string
        required: true
      - in: path
        name: role_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - members
      security:
      - sessionCookie: []
      responses:
        '204':
          description: No response body
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
    delete:
      operationId: members_roles_remove
      description: 'Quita el rol al miembro. 403: como al asignarlo. 404: el miembro
        no tiene ese rol, o no son de la organización. 409 `LAST_OWNER`: sería el
        último Owner activo, o la organización no tiene rol Owner.'
      parameters:
      - in: path
        name: membership_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: org_slug
        schema:
          type: string
        required: true
      - in: path
        name: role_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - members
      security:
      - sessionCookie: []
      responses:
        '204':
          description: No response body
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
  /api/v1/o/{org_slug}/permissions/:
    get:
      operationId: permissions_list
      description: Los permisos que se pueden conceder a un rol, ordenados por código.
      parameters:
      - in: path
        name: org_slug
        schema:
          type: string
        required: true
      tags:
      - roles
      security:
      - sessionCookie: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionCatalog'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/v1/o/{org_slug}/roles/:
    get:
      operationId: roles_list
      description: |-
        Los roles de la organización, con lo que concede cada uno y cuántos miembros lo tienen.
        Paginado por orden de creación (ADR-016). `POST` crea un rol propio (F2-29).
      parameters:
      - name: cursor
        in: query
        required: false
        description: El `next` de la página anterior. Sin él, la primera página.
        schema:
          type: string
      - name: limit
        in: query
        required: false
        description: Filas por página.
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - in: path
        name: org_slug
        schema:
          type: string
        required: true
      tags:
      - roles
      security:
      - sessionCookie: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoleList'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
    post:
      operationId: roles_create
      description: 'Crea un rol propio, vacío: sin permisos y sin miembros. El código
        lo genera el servidor. 409 `ROLE_NAME_TAKEN`: ya hay en la organización un
        rol cuyo nombre se lee igual (mayúsculas, espacios repetidos, formas Unicode).
        409 `LAST_OWNER`: la organización no tiene rol Owner y no admite ningún cambio.'
      parameters:
      - in: path
        name: org_slug
        schema:
          type: string
        required: true
      tags:
      - roles
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
        required: true
      security:
      - sessionCookie: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Código estable. Lo único que decide el cliente.
        message:
          type: string
          description: Texto seguro para mostrar.
        fields:
          type: object
          additionalProperties: {}
          description: 'Solo en VALIDATION_ERROR: errores por campo (`_`: generales).'
      required:
      - code
...
    MemberOrganization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
      required:
      - id
      - name
      - slug
...
    MemberUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
      required:
      - email
      - first_name
      - id
      - last_name
...
    PaginatedMemberList:
      type: object
      required:
      - results
      - next
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        next:
          type: string
          nullable: true
          description: Cursor de la página siguiente; `null` en la última.
    PaginatedRoleList:
      type: object
      required:
      - results
      - next
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        next:
          type: string
          nullable: true
          description: Cursor de la página siguiente; `null` en la última.
...
    PermissionCatalog:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
      required:
      - results
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        is_system:
          type: boolean
          description: Nació de una plantilla del producto.
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/RoleGrant'
        members:
          type: integer
          description: Membresías con este rol, en cualquier estado.
        editable:
          type: boolean
          description: 'Falso para el rol Owner y para un rol que tiene asignado quien
            pregunta: la API rechaza cualquier cambio de sus concesiones. Es una ayuda
            para la interfaz, no una autorización: cada escritura aplica sus reglas.'
      required:
      - code
      - description
      - editable
      - id
      - is_system
      - members
      - name
      - permissions
    RoleCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 255
      required:
      - name
...
    RoleName:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
      required:
      - code
      - name
...
    SelfContext:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/MemberUser'
        organization:
          $ref: '#/components/schemas/MemberOrganization'
        membership_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleName'
          description: 'Solo para mostrar: no autorizan nada.'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Grant'
      required:
      - membership_id
      - organization
      - permissions
      - roles
      - user
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: crm_session
      description: Cookie de sesión (`__Host-crm_session` en producción). Los métodos
        no seguros llevan además la cabecera `X-CSRFToken`.